% Auto-generated: do not edit by hand
\name{''Editor}

\alias{''Editor}

\title{Editor component}

\description{

}

\usage{
''Editor(id=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

