% Auto-generated: do not edit by hand
\name{''LanrzipSelectPopup}

\alias{''LanrzipSelectPopup}

\title{LanrzipSelectPopup component}

\description{

}

\usage{
''LanrzipSelectPopup(children=NULL, id=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

