% Auto-generated: do not edit by hand
\name{''LanrzipSlick}

\alias{''LanrzipSlick}

\title{LanrzipSlick component}

\description{

}

\usage{
''LanrzipSlick(children=NULL, id=NULL, accessibility=NULL,
adaptiveHeight=NULL, arrows=NULL, asNavFor=NULL,
autoplay=NULL, autoplaySpeed=NULL, centerMode=NULL,
centerPadding=NULL, className=NULL, dots=NULL,
dotsClass=NULL, draggable=NULL, easing=NULL, fade=NULL,
focusOnSelect=NULL, infinite=NULL, initialSlide=NULL,
lazyLoad=NULL, pauseOnDotsHover=NULL, pauseOnFocus=NULL,
pauseOnHover=NULL, responsive=NULL, rows=NULL, rtl=NULL,
slide=NULL, slidesPerRow=NULL, slidesToScroll=NULL,
slidesToShow=NULL, speed=NULL, swipe=NULL,
swipeToSlide=NULL, touchMove=NULL, touchThreshold=NULL,
useCSS=NULL, useTransform=NULL, variableWidth=NULL,
vertical=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{accessibility}{Logical. }

\item{adaptiveHeight}{Logical. }

\item{arrows}{Logical. }

\item{asNavFor}{Named list. }

\item{autoplay}{Logical. }

\item{autoplaySpeed}{Numeric. }

\item{centerMode}{Logical. }

\item{centerPadding}{Character. }

\item{className}{Character. }

\item{dots}{Logical. }

\item{dotsClass}{Character. }

\item{draggable}{Logical. }

\item{easing}{Character. }

\item{fade}{Logical. }

\item{focusOnSelect}{Logical. }

\item{infinite}{Logical. }

\item{initialSlide}{Numeric. }

\item{lazyLoad}{A value equal to: 'ondemand', 'progressive'. }

\item{pauseOnDotsHover}{Logical. }

\item{pauseOnFocus}{Logical. }

\item{pauseOnHover}{Logical. }

\item{responsive}{Unnamed list. }

\item{rows}{Numeric. }

\item{rtl}{Logical. }

\item{slide}{Character. }

\item{slidesPerRow}{Numeric. }

\item{slidesToScroll}{Numeric. }

\item{slidesToShow}{Numeric. }

\item{speed}{Numeric. }

\item{swipe}{Logical. }

\item{swipeToSlide}{Logical. }

\item{touchMove}{Logical. }

\item{touchThreshold}{Numeric. }

\item{useCSS}{Logical. }

\item{useTransform}{Logical. }

\item{variableWidth}{Logical. }

\item{vertical}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

